# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  models.CloudPrivateIp:
    properties:
      assigned_to_id:
        type: integer
      cloud_private_ip_id:
        type: integer
      created_at:
        type: string
      created_by_user_id:
        type: integer
      description:
        type: string
      ipv4_assignment:
        type: string
      ipv6_assignment:
        type: string
    type: object
  models.CloudPublicIp:
    properties:
      cloud_public_ip_id:
        type: integer
      cloud_vm_name:
        type: string
      created_at:
        type: string
      created_by_user_id:
        type: integer
      customer_location_id:
        type: integer
      deleted_at:
        type: string
      deleted_by_user_id:
        type: integer
      ip_address:
        type: string
      post_forward_ip:
        type: string
      updated_at:
        type: string
      updated_by_user_id:
        type: integer
    type: object
...
  models.CreateCloudPrivateIp:
    properties:
      created_by_user_id:
        type: integer
      description:
        type: string
      ipv4_assignment:
        type: string
      ipv6_assignment:
        type: string
    type: object
  models.CreateCloudPublicIp:
    properties:
      cloud_vm_name:
        type: string
      created_by_user_id:
        type: integer
      customer_location_id:
        type: integer
      ip_address:
        type: string
      post_forward_ip:
        type: string
    type: object
...
  models.CreateCustomerLan:
    properties:
      backup_date:
        type: string
      created_by_user_id:
        type: integer
      customer_location_id:
        type: integer
      equipment_installed:
        type: string
      extra:
        type: string
      firewall_ipv4_lan_address:
        type: string
      firewall_ipv6_lan_address:
        type: string
      gateway_type:
        type: string
      lan_notes:
        type: string
      management_ip_address:
        type: string
      management_notes:
        type: string
      management_server:
        type: string
      network_on_site:
        type: string
      number_of_internet_connection:
        type: integer
      on_site_backup_server_or_nas_ip_address:
        type: string
      on_site_backup_server_or_nas_type:
        type: string
      print_or_scanner_ip_address:
        type: string
      print_or_scanner_password:
        type: string
      print_or_scanner_type:
        type: string
      print_or_scanner_user_name:
        type: string
      private_ip_assignment_id:
        type: integer
      scan_to_email_email_or_password:
        type: string
      scan_to_email_smtp_server_port:
        type: string
      scan_to_folder_location:
        type: string
      scan_to_folder_username_or_password:
        type: string
      version_of_last_backup:
        type: string
    type: object
...
  models.CreateInternetProvider:
    properties:
      account_or_pin:
        type: string
      name:
        type: string
      other:
        type: string
      primary_dns:
        type: string
      secondary_dns:
        type: string
      speed:
        type: string
    type: object
...
  models.CreateSwitches:
    properties:
      switch_brand_or_model:
        type: string
      switch_credentials:
        type: string
      switch_image_links:
        type: string
      switch_install_date:
        type: string
      switch_ip_address:
        type: string
      switch_manage:
        type: string
      switch_notes:
        type: string
    type: object
...
  models.CreateWanConfig:
    properties:
      comment:
        type: string
      firewall:
        type: string
      gateway:
        type: string
      ip_range:
        type: string
      ip_version:
        type: string
      sub_net_mask:
        type: string
    type: object
  models.CreateWireless:
    properties:
      wireless_admin_password:
        type: string
      wireless_admin_username:
        type: string
      wireless_connection_type:
        type: string
      wireless_ip_address:
        type: string
      wireless_notes:
        type: string
      wireless_password:
        type: string
      wireless_ssid:
        type: string
      wireless_unit:
        type: string
    type: object
...
  models.CustomerLan:
    properties:
      backup_date:
        type: string
      cloud_private_ip:
        $ref: '#/definitions/models.CloudPrivateIp'
      created_at:
        type: string
      created_by_user_id:
        type: integer
      customer_lan_id:
        type: integer
      customer_location_id:
        type: integer
      deleted_at:
        type: string
      deleted_by_user_id:
        type: integer
      equipment_installed:
        type: string
      extra:
        type: string
      firewall_ipv4_lan_address:
        type: string
      firewall_ipv6_lan_address:
        type: string
      gateway_type:
        type: string
      internet_provider1:
        $ref: '#/definitions/models.InternetProvider'
      internet_provider1_id:
        type: integer
      internet_provider2:
        $ref: '#/definitions/models.InternetProvider'
      internet_provider2_id:
        type: integer
      lan_notes:
        type: string
      management_ip_address:
        type: string
      management_notes:
        type: string
      management_server:
        type: string
      network_on_site:
        type: string
      number_of_internet_connection:
        type: integer
      on_site_backup_server_or_nas_ip_address:
        type: string
      on_site_backup_server_or_nas_type:
        type: string
      print_or_scanner_ip_address:
        type: string
      print_or_scanner_password:
        type: string
      print_or_scanner_type:
        type: string
      print_or_scanner_user_name:
        type: string
      private_ip_assignment_id:
        type: integer
      scan_to_email_email_or_password:
        type: string
      scan_to_email_smtp_server_port:
        type: string
      scan_to_folder_location:
        type: string
      scan_to_folder_username_or_password:
        type: string
      switch_brand_or_model:
        type: string
      switch_credentials:
        type: string
      switch_image_links:
        type: string
      switch_install_date:
        type: string
      switch_ip_address:
        type: string
      switch_manage:
        type: string
      switch_notes:
        type: string
      updated_at:
        type: string
      updated_by_user_id:
        type: integer
      version_of_last_backup:
        type: string
      wireless_admin_password:
        type: string
      wireless_admin_username:
        type: string
      wireless_connection_type:
        type: string
      wireless_ip_address:
        type: string
      wireless_notes:
        type: string
      wireless_password:
        type: string
      wireless_ssid:
        type: string
      wireless_unit:
        type: string
    type: object
...
  models.InternetProvider:
    properties:
      account_or_pin:
        type: string
      name:
        type: string
      other:
        type: string
      primary_dns:
        type: string
      secondary_dns:
        type: string
      speed:
        type: string
      wan_config_ipv4:
        $ref: '#/definitions/models.WanConfig'
      wan_config_ipv4_id:
        type: integer
      wan_config_ipv6:
        $ref: '#/definitions/models.WanConfig'
      wan_config_ipv6_id:
        type: integer
    type: object
...
  models.UpdateCloudPrivateIp:
    properties:
      assigned_to_id:
        type: integer
      description:
        type: string
      ipv4_assignment:
        type: string
      ipv6_assignment:
        type: string
    type: object
  models.UpdateCloudPublicIp:
    properties:
      cloud_vm_name:
        type: string
      customer_location_id:
        type: integer
      ip_address:
        type: string
      post_forward_ip:
        type: string
      updated_by_user_id:
        type: integer
    type: object
...
  models.UpdateCustomerLan:
    properties:
      backup_date:
        type: string
      created_by_user_id:
        type: integer
      customer_location_id:
        type: integer
      equipment_installed:
        type: string
      extra:
        type: string
      firewall_ipv4_lan_address:
        type: string
      firewall_ipv6_lan_address:
        type: string
      gateway_type:
        type: string
      lan_notes:
        type: string
      management_ip_address:
        type: string
      management_notes:
        type: string
      management_server:
        type: string
      network_on_site:
        type: string
      number_of_internet_connection:
        type: integer
      on_site_backup_server_or_nas_ip_address:
        type: string
      on_site_backup_server_or_nas_type:
        type: string
      print_or_scanner_ip_address:
        type: string
      print_or_scanner_password:
        type: string
      print_or_scanner_type:
        type: string
      print_or_scanner_user_name:
        type: string
      private_ip_assignment_id:
        type: integer
      scan_to_email_email_or_password:
        type: string
      scan_to_email_smtp_server_port:
        type: string
      scan_to_folder_location:
        type: string
      scan_to_folder_username_or_password:
        type: string
      version_of_last_backup:
        type: string
    type: object
...
  models.UpdateInternetProvider:
    properties:
      account_or_pin:
        type: string
      name:
        type: string
      other:
        type: string
      primary_dns:
        type: string
      secondary_dns:
        type: string
      speed:
        type: string
    type: object
...
  models.UpdateSwitches:
    properties:
      switch_brand_or_model:
        type: string
      switch_credentials:
        type: string
      switch_image_links:
        type: string
      switch_install_date:
        type: string
      switch_ip_address:
        type: string
      switch_manage:
        type: string
      switch_notes:
        type: string
    type: object
  models.UpdateWanConfig:
    properties:
      comment:
        type: string
      firewall:
        type: string
      gateway:
        type: string
      ip_range:
        type: string
      ip_version:
        type: string
      sub_net_mask:
        type: string
    type: object
  models.UpdateWireless:
    properties:
      wireless_admin_password:
        type: string
      wireless_admin_username:
        type: string
      wireless_connection_type:
        type: string
      wireless_ip_address:
        type: string
      wireless_notes:
        type: string
      wireless_password:
        type: string
      wireless_ssid:
        type: string
      wireless_unit:
        type: string
    type: object
...
  models.WanConfig:
    properties:
      comment:
        type: string
      firewall:
        type: string
      gateway:
        type: string
      ip_range:
        type: string
      ip_version:
        type: string
      sub_net_mask:
        type: string
      wan_config_id:
        type: integer
    type: object
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: This is a sample server .
  termsOfService: http://swagger.io/terms/
  title: Network Management Solutions API
  version: "1.0"
paths:
  /cloud-private-ips:
    delete:
      description: Delete a Cloud Private IP
      parameters:
      - description: Cloud Private IP ID
        in: query
        name: cloud_private_ip_id
        type: string
      - description: USER ID
        in: header
        name: USER_ID
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.CloudPrivateIp'
      summary: Delete Cloud Private IP
      tags:
      - Cloud Private IP
    get:
      description: Retrieve  Cloud Private IP
      parameters:
      - description: Page
        in: query
        name: page
        type: string
      - description: Cloud Private IP ID
        in: query
        name: cloud_private_ip_id
        type: string
      - description: Select Column
        in: query
        name: select_column
        type: string
      - description: Append Select
        in: query
        name: append_select
        type: string
      - description: Set Limit
        in: query
        name: set_limit
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.CloudPrivateIp'
      summary: Retrieve  Cloud Private IP
      tags:
      - Cloud Private IP
    patch:
      consumes:
      - application/json
      description: Update a Cloud Private IP
      parameters:
      - description: CloudPrivateIp Data
        in: body
        name: CloudPrivateIP
        required: true
        schema:
          $ref: '#/definitions/models.UpdateCloudPrivateIp'
      - description: Cloud Private IP ID
        in: query
        name: cloud_private_ip_id
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.CloudPrivateIp'
      summary: Update Cloud Private IP
      tags:
      - Cloud Private IP
    post:
      consumes:
      - application/json
      description: Add a Cloud Private IP
      parameters:
      - description: CloudPrivateIp Data
        in: body
        name: CloudPrivateIP
        required: true
        schema:
          $ref: '#/definitions/models.CreateCloudPrivateIp'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.CloudPrivateIp'
      summary: Create Cloud Private IP
      tags:
      - Cloud Private IP
  /cloud-public-ips:
    delete:
      description: Delete a Cloud Public IP
      parameters:
      - description: Cloud Public IP ID
        in: query
        name: cloud_public_ip_id
        type: string
      - description: USER ID
        in: header
        name: USER_ID
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.CloudPublicIp'
      summary: Delete Cloud Public IP
      tags:
      - Cloud Public IP
    get:
      description: Retrieve a Cloud Public IP
      parameters:
      - description: Cloud Public IP ID
        in: query
        name: cloud_public_ip_id
        type: string
      - description: Ip Address
        in: query
        name: ip_address
        type: string
      - description: Location
        in: query
        name: location
        type: string
      - description: Cloud VM Name
        in: query
        name: cloud_vm_name
        type: string
      - description: Page
        in: query
        name: page
        type: integer
      - description: Order By
        in: query
        name: order_by
        type: string
      - description: Post Forward Ip
        in: query
        name: post_forward_ip
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.CloudPublicIp'
      summary: Retrieve Cloud Public IP
      tags:
      - Cloud Public IP
    patch:
      consumes:
      - application/json
      description: Update a Cloud Public IP
      parameters:
      - description: CloudPublicIp Data
        in: body
        name: CloudPublicIp
        required: true
        schema:
          $ref: '#/definitions/models.UpdateCloudPublicIp'
      - description: Cloud Public IP ID
        in: query
        name: cloud_public_ip_id
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.CloudPublicIp'
      summary: Update Cloud Public IP
      tags:
      - Cloud Public IP
    post:
      consumes:
      - application/json
      description: Add a Cloud Public IP
      parameters:
      - description: CloudPublicIp Data
        in: body
        name: CloudPublicIp
        required: true
        schema:
          $ref: '#/definitions/models.CreateCloudPublicIp'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.CloudPublicIp'
      summary: Create Cloud Public IP
      tags:
      - Cloud Public IP
  /customer-lans:
    delete:
      description: Delete Customer Lan
      parameters:
      - description: Cloud Public IP ID
        in: query
        name: customer_lan_id
        type: string
      - description: USER ID
        in: header
        name: User_id
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.CustomerLan'
      summary: Delete Customer Lan
      tags:
      - Customer Lan
    get:
      description: Retrieve Customer Lan
      parameters:
      - description: Customer Location
        in: query
        name: customer_location
        type: string
      - description: Customer Lan Network On Site
        in: query
        name: customer_lan_network_on_site
        type: string
      - description: Assign To
        in: query
        name: assign_to
        type: string
      - description: Page
        in: query
        name: page
        type: integer
      - description: Order By
        in: query
        name: order_by
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.CustomerLan'
      summary: Retrieve Customer Lan
      tags:
      - Customer Lan
    patch:
      consumes:
      - application/json
      description: Update Customer Lan
      parameters:
      - description: CustomerLan Data
        in: body
        name: CustomerLan
        required: true
        schema:
          allOf:
          - $ref: '#/definitions/models.UpdateCustomerLan'
          - properties:
              internet_provider1:
                allOf:
                - $ref: '#/definitions/models.UpdateInternetProvider'
                - properties:
                    wan_config_ipv4:
                      $ref: '#/definitions/models.UpdateWanConfig'
                    wan_config_ipv6:
                      $ref: '#/definitions/models.UpdateWanConfig'
                  type: object
              internet_provider2:
                allOf:
                - $ref: '#/definitions/models.UpdateInternetProvider'
                - properties:
                    wan_config_ipv4:
                      $ref: '#/definitions/models.UpdateWanConfig'
                    wan_config_ipv6:
                      $ref: '#/definitions/models.UpdateWanConfig'
                  type: object
              switches:
                $ref: '#/definitions/models.UpdateSwitches'
              wireless:
                $ref: '#/definitions/models.UpdateWireless'
            type: object
      - description: Customer Lan ID
        in: query
        name: customer_lan_id
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.CustomerLan'
      summary: Update Customer Lan
      tags:
      - Customer Lan
    post:
      consumes:
      - application/json
      description: Add a Customer Lan
      parameters:
      - description: CustomerLan Data
        in: body
        name: CustomerLan
        required: true
        schema:
          allOf:
          - $ref: '#/definitions/models.CreateCustomerLan'
          - properties:
              internet_provider1:
                allOf:
                - $ref: '#/definitions/models.CreateInternetProvider'
                - properties:
                    wan_config_ipv4:
                      $ref: '#/definitions/models.CreateWanConfig'
                    wan_config_ipv6:
                      $ref: '#/definitions/models.CreateWanConfig'
                  type: object
              internet_provider2:
                allOf:
                - $ref: '#/definitions/models.CreateInternetProvider'
                - properties:
                    wan_config_ipv4:
                      $ref: '#/definitions/models.CreateWanConfig'
                    wan_config_ipv6:
                      $ref: '#/definitions/models.CreateWanConfig'
                  type: object
              switches:
                $ref: '#/definitions/models.CreateSwitches'
              wireless:
                $ref: '#/definitions/models.CreateWireless'
            type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.CustomerLan'
      summary: Create Customer Lan
      tags:
      - Customer Lan
...
schemes:
- http
swagger: "2.0"